# Internship_API/src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Internship_API
  description: This is the Internship API documentation. My goal is to learn best practices about REST API Design and Architecture.
  version: 1.0.0
servers:
  - url: http://localhost:8081/v2
paths:
...
  /route/get:
    get:
      tags:
        - Route
      summary: Get all routes
      operationId: getRoutes
      security:
        - internship_api_security_scheme:
            - read_access
      parameters:
        - name: search
          in: query
          description: Filter for company prices
          required: false
          schema:
            $ref: "#/components/schemas/RouteSearchObject"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/List"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/RouteDTO"
  /route/getById/{id}:
    get:
      tags:
        - Route
      summary: Get route by ID
      operationId: getRouteById
      security:
        - internship_api_security_scheme:
            - read_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDTO"
  /route/getTopClients:
    get:
      tags:
        - Route
      summary: Get clients by route count
      operationId: getTopClients
      security:
        - internship_api_security_scheme:
            - read_access
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteClientCountDTO"
  /route/getAmountForReport:
    post:
      tags:
        - Route
      summary: Get report amount
      operationId: getAmountForReport
      security:
        - internship_api_security_scheme: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneralReportRequest"
      responses:
        "200":
          description: Report values
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: number
                  format: double
  /route/save:
    post:
      tags:
        - Route
      summary: Save a new route
      operationId: saveRoute
      security:
        - internship_api_security_scheme:
            - write_access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInsertRequest"
      responses:
        "201":
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDTO"
  /route/update/{id}:
    put:
      tags:
        - Route
      summary: Update route details by ID
      operationId: updateRoute
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          description: The ID of the route to update
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdateRequest"
      responses:
        "200":
          description: Route updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /route/updateFinish/{id}:
    put:
      tags:
        - Route
      summary: Update finish route
      operationId: updateRouteFinish
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route finished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDTO"
        "404":
          description: The Route with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /route/updatePayment/{id}:
    put:
      tags:
        - Route
      summary: Update payment route
      operationId: updateRoutePayment
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDTO"
        "404":
          description: The Route with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /route/delete/{id}:
    delete:
      tags:
        - Route
      summary: Delete route
      operationId: deleteRoute
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDTO"
  #Vehicle paths
...
components:
  securitySchemes:
    internship_api_security_scheme:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://keycloak:8080/realms/Internship_API/protocol/openid-connect/auth
          tokenUrl: http://keycloak:8080/realms/Internship_API/protocol/openid-connect/token
          scopes:
            read_access: read data
            write_access: modify data
  schemas:
    List:
      type: object
      properties:
        count:
          type: integer
    #User data objecets
    UserDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        surname:
          type: string
        username:
          type: string
        email:
          type: string
        telephoneNumber:
          type: string
        gender:
          type: string
        registrationDate:
          type: string
          format: date-time
...
    ClientDTO:
      type: object
      properties:
        id:
          type: integer
        image:
          type: string
          format: byte
        user:
          $ref: "#/components/schemas/UserDTO"
...
    DriverDTO:
      type: object
      properties:
        id:
          type: integer
        numberOfClientsAmount:
          type: integer
        numberOfHoursAmount:
          type: integer
        user:
          $ref: "#/components/schemas/UserDTO"
...
    RouteDTO:
      type: object
      properties:
        id:
          type: integer
        sourcePointLat:
          type: number
          format: double
        sourcePointLon:
          type: number
          format: double
        destinationPointLat:
          type: number
          format: double
        destinationPointLon:
          type: number
          format: double
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        duration:
          type: integer
        numberOfKilometers:
          type: number
          format: double
        fullPrice:
          type: number
          format: double
        paid:
          type: boolean
        status:
          type: string
        client:
          $ref: "#/components/schemas/ClientDTO"
        driver:
          $ref: "#/components/schemas/DriverDTO"
    RouteInsertRequest:
      type: object
      required:
        - sourcePointLat
        - sourcePointLon
        - destinationPointLat
        - destinationPointLon
        - client_id
        - driver_id
      properties:
        sourcePointLat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: "Latitude must be between -90 and 90 degrees"
        sourcePointLon:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: "Longitude must be between -180 and 180 degrees"
        destinationPointLat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: "Latitude must be between -90 and 90 degrees"
        destinationPointLon:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: "Longitude must be between -180 and 180 degrees"
        client_id:
          type: integer
        driver_id:
          type: integer
    RouteUpdateRequest:
      type: object
      properties:
        client_id:
          type: integer
        driver_id:
          type: integer
    RouteSearchObject:
      type: object
      properties:
        status:
          type: string
        client_id:
          type: integer
        driver_id:
          type: integer
        user_id:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
    GeneralReportRequest:
      type: object
      required:
        - month
        - year
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 2024
    RouteClientCountDTO:
      type: object
      properties:
        client:
          type: string
        count:
          type: int
     



    #Vehicle data objects
...